+++ types/components/index.d.ts
@@ -1,8 +1,0 @@
-export * from './overlays';
-export * from './inputs';
-export * from './containers';
-export * from './router-outlet';
-export * from './back-button';
-export * from './tabs';
-export * from './icon';
-//# sourceMappingURL=index.d.ts.map