+++ types/components/tabs/index.d.ts
@@ -1,5 +1,0 @@
-export { IonTabs } from './tabs';
-export { IonTabBar } from './tab-bar';
-export * from './tab';
-export * from './tab-button';
-//# sourceMappingURL=index.d.ts.map