+++ types/components/tabs/tab-bar.d.ts
@@ -1,6 +1,0 @@
-import { FunctionalComponent } from 'vue';
-import { JSX } from '@ionic/core';
-export declare const tabBarRef: import("vue").Ref<HTMLIonTabBarElement | undefined>;
-export declare const setActiveTab: (value?: string | undefined) => void;
-export declare const IonTabBar: FunctionalComponent<JSX.IonTabBar>;
-//# sourceMappingURL=tab-bar.d.ts.map