+++ types/components/tabs/tab-button.d.ts
@@ -1,4 +1,0 @@
-import { FunctionalComponent } from 'vue';
-import { JSX } from '@ionic/core';
-export declare const IonTabButton: FunctionalComponent<JSX.IonTabButton>;
-//# sourceMappingURL=tab-button.d.ts.map