+++ types/components/tabs/tab.d.ts
@@ -1,4 +1,0 @@
-import { FunctionalComponent } from 'vue';
-import { JSX } from '@ionic/core';
-export declare const IonTab: FunctionalComponent<JSX.IonTab>;
-//# sourceMappingURL=tab.d.ts.map