+++ types/components/tabs/tabs.d.ts
@@ -1,4 +1,0 @@
-import { FunctionalComponent, VNode } from 'vue';
-export declare const tabNodes: import("vue").Ref<Map<string, VNode<import("vue").RendererNode, import("vue").RendererElement>>>;
-export declare const IonTabs: FunctionalComponent;
-//# sourceMappingURL=tabs.d.ts.map