+++ types/utils/index.d.ts
@@ -1,4 +1,0 @@
-export * from './defineOverlay';
-export * from './defineInput';
-export * from './defineContainer';
-//# sourceMappingURL=index.d.ts.map